+++ src/Akeneo/Pim/Enrichment/Product/back/Test/Specification/API/Command/UpsertProductCommandSpec.php
@@ -9,6 +9,7 @@ use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\AddMultiSelectValue;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\ClearValue;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetBooleanValue;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetDateValue;
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetFamily;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetMetricValue;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetNumberValue;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetTextareaValue;
@@ -42,89 +43,18 @@ class UpsertProductCommandSpec extends ObjectBehavior
         $this->shouldHaveType(UpsertProductCommand::class);
     }
 
-    function it_can_be_constructed_with_a_set_text_value_intent()
+    function it_can_be_constructed_with_value_intents()
     {
-        $valueUserIntents = [new SetTextValue('name', null, null, 'foo')];
-        $this->beConstructedWith(
-            1,
-            'identifier1',
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            $valueUserIntents
-        );
-        $this->userId()->shouldReturn(1);
-        $this->productIdentifier()->shouldReturn('identifier1');
-        $this->valueUserIntents()->shouldReturn($valueUserIntents);
-    }
-
-    function it_can_be_constructed_with_a_set_number_value_intent()
-    {
-        $valueUserIntents = [new SetNumberValue('name', null, null, 10)];
-        $this->beConstructedWith(
-            1,
-            'identifier1',
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            $valueUserIntents
-        );
-        $this->userId()->shouldReturn(1);
-        $this->productIdentifier()->shouldReturn('identifier1');
-        $this->valueUserIntents()->shouldReturn($valueUserIntents);
-    }
-
-    function it_can_be_constructed_with_a_set_metric_value_intent()
-    {
-        $valueUserIntents = [new SetMetricValue('power', null, null, '100', 'KILOWATT')];
-        $this->beConstructedWith(
-            1,
-            'identifier1',
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            $valueUserIntents
-        );
-        $this->userId()->shouldReturn(1);
-        $this->productIdentifier()->shouldReturn('identifier1');
-        $this->valueUserIntents()->shouldReturn($valueUserIntents);
-    }
-
-    function it_can_be_constructed_with_a_set_textarea_value_intent()
-    {
-        $valueUserIntents = [new SetTextareaValue('name', null, null, "<p><span style=\"font-weight: bold;\">title</span></p><p>text</p>")];
-        $this->beConstructedWith(
-            1,
-            'identifier1',
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            $valueUserIntents
-        );
-        $this->userId()->shouldReturn(1);
-        $this->productIdentifier()->shouldReturn('identifier1');
-        $this->valueUserIntents()->shouldReturn($valueUserIntents);
-    }
-
-    function it_can_be_constructed_with_a_clear_value_intent()
-    {
-        $valueUserIntents = [new ClearValue('name', null, null)];
+        $valueUserIntents = [
+            new SetTextValue('name', null, null, 'foo'),
+            new SetNumberValue('name', null, null, '10'),
+            new SetMetricValue('power', null, null, '100', 'KILOWATT'),
+            new SetTextareaValue('name', null, null, "<p><span style=\"font-weight: bold;\">title</span></p><p>text</p>"),
+            new ClearValue('name', null, null),
+            new SetBooleanValue('name', null, null, true),
+            new SetDateValue('name', null, null, new \DateTime("2022-03-04T09:35:24+00:00")),
+            new AddMultiSelectValue('name', null, null, ['optionA']),
+        ];
         $this->beConstructedWith(
             1,
             'identifier1',
@@ -159,63 +89,44 @@ class UpsertProductCommandSpec extends ObjectBehavior
         $this->shouldThrow(\InvalidArgumentException::class)->duringInstantiation();
     }
 
-    function it_can_be_constructed_with_a_set_boolean_value_intent()
+    function it_can_be_constructed_with_field_user_intents()
     {
-        $valueUserIntents = [new SetBooleanValue('name', null, null, true)];
+        $familyUserIntent = new SetFamily('accessories');
         $this->beConstructedWith(
             1,
             'identifier1',
             null,
+            $familyUserIntent,
             null,
             null,
             null,
             null,
             null,
-            null,
-            $valueUserIntents
+            []
         );
         $this->userId()->shouldReturn(1);
         $this->productIdentifier()->shouldReturn('identifier1');
-        $this->valueUserIntents()->shouldReturn($valueUserIntents);
+        $this->familyUserIntent()->shouldReturn($familyUserIntent);
+        $this->valueUserIntents()->shouldReturn([]);
     }
 
-    function it_can_be_constructed_with_a_set_date_value_intent()
+    function it_can_be_constructed_from_a_collection_of_user_intents()
     {
-        $valueUserIntents = [new SetDateValue('name', null, null, new \DateTime("2022-03-04T09:35:24+00:00"))];
-        $this->beConstructedWith(
-            1,
-            'identifier1',
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            $valueUserIntents
-        );
-        $this->userId()->shouldReturn(1);
-        $this->productIdentifier()->shouldReturn('identifier1');
-        $this->valueUserIntents()->shouldReturn($valueUserIntents);
-    }
+        $familyUserIntent = new SetFamily('accessories');
+        $setTextValue = new SetTextValue('name', null, null, 'foo');
+        $setNumberValue = new SetNumberValue('name', null, null, '10');
+        $setDateValue = new SetDateValue('name', null, null, new \DateTime("2022-03-04T09:35:24+00:00"));
+        $addMultiSelectValue = new AddMultiSelectValue('name', null, null, ['optionA']);
 
-    function it_can_be_constructed_with_a_add_multi_select_value_intent()
-    {
-        $valueUserIntents = [new AddMultiSelectValue('name', null, null, ['optionA'])];
-        $this->beConstructedWith(
-            1,
+        $this->beConstructedThrough('createFromCollection', [
+            10,
             'identifier1',
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null,
-            $valueUserIntents
-        );
-        $this->userId()->shouldReturn(1);
+            [$familyUserIntent, $setTextValue, $setNumberValue, $setDateValue, $addMultiSelectValue]
+        ]);
+
+        $this->userId()->shouldReturn(10);
         $this->productIdentifier()->shouldReturn('identifier1');
-        $this->valueUserIntents()->shouldReturn($valueUserIntents);
+        $this->familyUserIntent()->shouldReturn($familyUserIntent);
+        $this->valueUserIntents()->shouldReturn([$setTextValue, $setNumberValue, $setDateValue, $addMultiSelectValue]);
     }
 }