+++ tests/back/Pim/Enrichment/Integration/Storage/Sql/ElasticsearchProjection/GetElasticsearchProductModelProjectionIntegration.php
@@ -340,7 +340,7 @@ class GetElasticsearchProductModelProjectionIntegration extends TestCase
 
     private function getProductModelProjectionArray($code): array
     {
-        $productModelProjections = $this->getGetElasticsearchProductModelProjection()->fromProductModelCodes([$code]);
+        $productModelProjections = $this->getElasticsearchProductModelProjection()->fromProductModelCodes([$code]);
         if (!\is_array($productModelProjections)) {
             $productModelProjections = \iterator_to_array($productModelProjections);
         }
@@ -348,7 +348,7 @@ class GetElasticsearchProductModelProjectionIntegration extends TestCase
         return $productModelProjections[$code]->toArray();
     }
 
-    private function getGetElasticsearchProductModelProjection(): GetElasticsearchProductModelProjectionInterface
+    private function getElasticsearchProductModelProjection(): GetElasticsearchProductModelProjectionInterface
     {
         return $this->get('akeneo.pim.enrichment.product.query.get_elasticsearch_product_model_projection');
     }