+++ CHANGELOG.md
@@ -14,7 +14,9 @@
 - PIM-10292: Fix error 500 when role page contain a validation errors
 - PIM-10268: SKU filter is always shown in the product grid
 - PIM-10295: Fixed product categories disappearing on adjacent tab locale switch
+- PIM-10330: Fix wrong error message while importing boolean attribute value
 - PIM-10331: Fix error when using an association with quantities having an numeric code
+- PIM-10336: Fix product Export edition in error if no locale selected
 
 ## Improvements
 - PIM-10293: add batch-size option to pim:completness:calculate command